+++ TkInforHard/widgets/__init__.py
@@ -6,7 +6,7 @@ from TkInforHard.widgets.data import IHEmptyState, IHFilterBar, IHPagination, IH
 from TkInforHard.widgets.display import IHBadge, IHDivider, IHLogo, IHSectionHeader
 from TkInforHard.widgets.feedback import IHAlert, IHLoading, IHProgress, IHToast
 from TkInforHard.widgets.inputs import IHCombobox, IHDateInput, IHInput, IHSearchInput, IHTextArea
-from TkInforHard.widgets.navigation import IHBreadcrumb, IHMenuItem, IHSidebar, IHTopbar
+from TkInforHard.widgets.navigation import IHBreadcrumb, IHDrawerMenu, IHMenuItem, IHSidebar, IHTopbar
 
 __all__ = [
     "IHButton",
@@ -28,6 +28,7 @@ __all__ = [
     "IHTopbar",
     "IHMenuItem",
     "IHBreadcrumb",
+    "IHDrawerMenu",
     "IHTable",
     "IHFilterBar",
     "IHPagination",
@@ -37,4 +38,3 @@ __all__ = [
     "IHDivider",
     "IHLogo",
 ]
-