+++ TkInforHard/widgets/navigation/__init__.py
@@ -1,9 +1,9 @@
 """Navigation components."""
 
 from TkInforHard.widgets.navigation.breadcrumb import IHBreadcrumb
+from TkInforHard.widgets.navigation.drawer_menu import IHDrawerMenu
 from TkInforHard.widgets.navigation.menu_item import IHMenuItem
 from TkInforHard.widgets.navigation.sidebar import IHSidebar
 from TkInforHard.widgets.navigation.topbar import IHTopbar
 
-__all__ = ["IHSidebar", "IHTopbar", "IHMenuItem", "IHBreadcrumb"]
-
+__all__ = ["IHSidebar", "IHTopbar", "IHMenuItem", "IHBreadcrumb", "IHDrawerMenu"]